+++ RELEASE_NOTES.md
@@ -1,8 +1,8 @@
 #### 5.1.0
-* __It happened. Farkle can create grammars without the need of GOLD Parser. Farkle is now a parsing library on its own.__
+* It happened. Farkle can create grammars without the need of GOLD Parser. __Farkle is now a parsing library on its own.__
 * Move the `CharStream` type in the `Farkle.IO` namespace.
 * Add methods to parse text from .NET `TextReader`s. They should be preferred over parsing .NET `Stream`s because the latter are supposed to contain binary data, not text.
-* The `Farkle.CSharp` namespace is no longer used. C# users just have to use `Farkle` to get their extension methods, unless they are writing their own post-processors for GOLD Parser grammars, where they have to use `Farkle.PostProcessor.CSharp`.
+* The `Farkle.CSharp` namespace is no longer required. C# users just have to use `Farkle` to get their extension methods, unless they are writing their own post-processors for GOLD Parser grammars, where they have to use `Farkle.PostProcessor.CSharp`.
 * __Breaking change:__ Farkle.Tools.MSBuild was upgraded to .NET Core 3.1. Nothing significant changed though, which means those who still use .NET Core 2.1 can stay in a previous version.
 * __Breaking change:__ Reading grammars from EGT files now raises an exception. [Rationale](https://github.com/teo-tsirpanis/Farkle/commit/84b96ccc85753e4deebffcf561cc89edb1e252bc)
 * __Breaking change:__ Some utility functions that had nothing to do with parsing were either removed or made internal.