+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-#### 5.1.0
+#### 5.1.0 - 31-01-2019
 * It happened. Farkle can create grammars without the need of GOLD Parser. __Farkle is now a parsing library on its own.__
 * Move the `CharStream` type in the `Farkle.IO` namespace.
 * Add methods to parse text from .NET `TextReader`s. They should be preferred over parsing .NET `Stream`s because the latter are supposed to contain binary data, not text.